Create a new app to store users
Create a new app called users for storing our site's users and
define the CustomUser model that replaces the standard User model.
Configure settings for this app and make migrations.
Configure settings for both local and production development by
creating the settings folder and three files inside it: base.py
for base settings and local.py and prod.py for the local and
production development respectfully.
Configure the docker-compose file.